+++ mitra-cli/src/main.rs
@@ -53,6 +53,8 @@ async fn main() {
                 SubCommand::UpdateCurrentBlock(cmd) => cmd.execute(&config, db_client).await.unwrap(),
                 SubCommand::ResetSubscriptions(cmd) => cmd.execute(&config, db_client).await.unwrap(),
                 SubCommand::CreateMoneroWallet(cmd) => cmd.execute(&config).await.unwrap(),
+                SubCommand::CreateMoneroSignature(cmd) => cmd.execute(&config).await.unwrap(),
+                SubCommand::VerifyMoneroSignature(cmd) => cmd.execute(&config).await.unwrap(),
                 SubCommand::CheckExpiredInvoice(cmd) => cmd.execute(&config, db_client).await.unwrap(),
                 _ => unreachable!(),
             };