+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Validate monero chain ID when reading configuration file.
 - Support managed database connections with TLS ([#34](https://codeberg.org/silverpill/mitra/pulls/34)).
 - Prevent re-use of EIP-4361 nonces.
+- Added `create-monero-signature` and `verify-monero-signature` commands.
 
 ### Changed